User break added to Io
It is implemented in IoMessage_locals_performOn_ because the SIGINT
querying happens there as well. It happens via a call to IoState_error_

User break now exists for all embedded languages.

Added svrSend callback to Io. You need to get the socket number from lua
and enter it as a parameter manually for now.

Added a call to praxis_putty.bat purely for my convenience. Remove it if
you wish.

Added forthGetState so the programmer can know what state Forth is in as
a program is entered.

Changed PraxisServer::Send so it is no longer limited to 65536 bytes.
This was going to bite eventually, and it did when I tried testing
infinite loop breaking in Io via loop("foo" println) and crashed praxis
when breaking it, because when it breaks it tries to send a trace which
far exceeds 65536 bytes all at once. It is also why I added svrSend to Io.