Removed a mistaken lua_pop call.
The original idea of this lua_pop call was to pop
the results of luaCall. This was never necessary. In
fact it is harmful because it causes onerror to no longer work.

I noticed a problem when debug.traceback sometimes causes a
null dereference deep in Lua.

I shouldn't be popping items I didn't push.