The text shader now works the same way as it does on ShaderToy.
Because I'm using the old OpenGL pipeline to draw the quad, the alpha
channel was having the side effect of altering the brightness of the
color channels, which is fine normally, but not when we need the raw data
from the other channels. Disabling GL_BLEND fixes it.

The next step is to assemble the string texture from a string,
instead of from the shader.