String -> Texture -> Shader experiment successful!
It is now possible to call glStringToTexture to
freely change the string the string shader uses!

The key breakthrough was when I realized that the hardcoded bytes
in the example weren't the bytes that were actually in the texture memory,
but instead were used together to be interpretted as a floating point
number, 3 bytes at a time. This also explains why the 4th byte couldn't be
used, because floating point can't handle the magnitude of a 4 byte integer.

Turned off gesture support by default so praxis will run on old windows
machines (anything earlier than Win7 - Vista, XP, 2000)