Turning the text shader into a usable editor
Added textshader_writebuffer which takes an "at" function and a length
so it can access the input string directly instead of a copy of it
being made first.

Minor adjustment to readChar in the shader because on some machines,
spaces (ASCII 32) were coming up as ASCII 31, or some fractional thing
in between.